requirements.txt
# django~=4.2
# djangorestframework
# django-filter
# djangorestframework-simplejwt
# drf-nested-routers
# pytest-django
# django-cors-headers
# markdown
# drf_yasg
# drf_spectacular
# whitenoise
# gunicorn
# python-dotenv
# django-countries
# django-extensions
# werkzeug
# pyOpenSSL
# rave-python
# python-decouple
# resend
# psycopg2-binary
# django-environ
# dj-rest-auth
# django-allauth
# six

asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
dj-rest-auth==6.0.0
Django==4.2.11
django-allauth==0.62.1
django-cors-headers==4.3.1
django-countries==7.6.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
drf-spectacular==0.27.2
drf-yasg==1.21.7
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Markdown==3.6
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.5.0
pycoin==0.92.20230326
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pypng==0.20220715.0
pytest==8.2.0
pytest-django==4.8.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
rave-python==1.4.0
referencing==0.35.1
requests==2.31.0
resend==0.8.0
rpds-py==0.18.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.2
whitenoise==6.6.0
qrcode